Modify options to add optional particle initial attributes
Particles attributes provided a single function to determine their
value. For particles born at the beginning of the simulation, testing
`t == 0` is sufficient to determine whether the particle has just
spawned, or is pre-existing. However, the ability to spawn further
particles during the simulation breaks that test: we can't test
whether a particle is new or pre-existing after time t=0.

As a step to offering a solution to this, particle attributes can have
an optional additional function, to specify the initial attribute
value only. If not specified, the original behaviour will be
retained. As yet, this is only an option change, the accompanying code
change follows.